svg/import/struct-dom-01-b-manual.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" id="svg-root" width="100%" height="100%" baseProfile="basic" onload="domTest(evt)" version="1.1" viewBox="0 0 480 360"><d:SVGTestCase xmlns:d="http://www.w3.org/2000/02/svg/testsuite/description/" author="VH" reviewer="SVGWG" status="accepted" template-version="1.4" testname="$RCSfile: struct-dom-01-b.svg,v $" version="$Revision: 1.8 $"><d:testDescription xmlns="http://www.w3.org/1999/xhtml" href="http://www.w3.org/TR/SVG11/struct.html#DOMInterfaces"><p>Verify the basic capability to handle the SVG DOM API.</p><p>The test is composed of a top level svg element with an &apos;onload&apos; event handler and a rect element. Both the svg and the rect elements have an identifier. The &apos;onload&apos; handler invokes SVG-specific DOM API methods which use these identifiers.</p><p>First, the handler gets the SVG element owner of the rect element and checks it has the expected identifier. Then, the handler accesses the coordinates of the rect element and uses them to build a &apos;shadow&apos; rectangle under the existing one. Finally, the &apos;shadow&apos; rectangle is created using the SVGSVGElement&apos;s createSVGRect method.</p></d:testDescription><d:operatorScript xmlns="http://www.w3.org/1999/xhtml"><p>Run the test. No interaction required.</p></d:operatorScript><d:passCriteria xmlns="http://www.w3.org/1999/xhtml"><p>The test passes if:</p><ul><li>The text &quot;This document&apos;s root identifier is: svg-root&quot; is shown.</li><li>A green rectangle with a black shadow is shown.</li></ul></d:passCriteria></d:SVGTestCase><title id="test-title">$RCSfile: struct-dom-01-b.svg,v $</title><defs><font-face font-family="SVGFreeSansASCII" unicode-range="U+0-7F"><font-face-src><font-face-uri xlink:href="../resources/SVGFreeSans.svg#ascii"/></font-face-src></font-face></defs><g id="test-body-content" font-family="SVGFreeSansASCII,sans-serif" font-size="18"><script type="text/ecmascript" xlink:href="../resources/testharness.js"/><script type="text/ecmascript"><![CDATA[
      function domTest(evt) {

      var svg_ns = "http://www.w3.org/2000/svg";

      // Get Document
      var target = evt.target;
      var doc = target.ownerDocument;

      //
      // Test that our rectangle is an SVGElement instance
      //
      var rect = doc.getElementById("rectId");
      var rootSVG = rect.ownerSVGElement;
      var rootId = rootSVG.getAttribute( "id" );

      // Insert a new text element to the DOM tree using the id
      var newText = doc.createElementNS(svg_ns, 'text');
      newText.setAttribute('x', '50');
      newText.setAttribute('y', '100');
      var message = "This document's root identifier is: " + rootId;
      var textContent = doc.createTextNode(message);
      newText.appendChild(textContent);
      rect.parentNode.appendChild(newText);

      //
      // Now, check that our rectangle is an instance of SVGRect by accessing
      // specific methods in order to get its x, y, width and height attributes.
      //
      var x = rect.x.baseVal.value; // SVGRect -&gt; SVGAnimatedLenght -&gt; SVGLength -&gt; long
      var y = rect.y.baseVal.value;
      var width = rect.width.baseVal.value;
      var height = rect.height.baseVal.value;

      //
      // Now, build a new SVGRect through the SVGSVGElement interface.
      //
      var newRect = doc.createElementNS(svg_ns, 'rect');

      //
      // Set the x, y, width and height of this element
      //
      newRect.x.baseVal.value = x + 10;
      newRect.y.baseVal.value = y + 10;
      newRect.setAttribute("width", width);
      newRect.setAttribute("height", height);

      //
      // Insert new element in DOM tree
      //
      rect.parentNode.insertBefore(newRect, rect);

      //
      // Check the pass criteria using the JS framework
      //
      test(function() {assert_equals(rootId, "svg-root")}, "Assert that the document's root identifier is 'svg-root'.");
      test(function() {
           assert_true(newRect instanceof SVGRectElement);
           assert_equals(newRect.getAttribute('x'), String(x+10));
           assert_equals(newRect.getAttribute('y'), String(y+10));
           assert_equals(newRect.getAttribute('width'), String(width));
           assert_equals(newRect.getAttribute('height'), String(height));
      }, "Assert that 'newRect' is replica (ignoring fill) of 'rect' with different x and y.");
      }

    ]]></script><rect id="rectId" width="50" height="50" x="40" y="150" fill="green"/></g><g font-family="SVGFreeSansASCII,sans-serif" font-size="32"><text id="revision" x="10" y="340" fill="#000" stroke="none">$Revision: 1.8 $</text></g><rect id="test-frame" width="478" height="358" x="1" y="1" fill="none" stroke="#000"/></svg>
